Mimic scroll docs, #169, #195, move all effects to using SpellStats

+++ src/main/java/com/hollingsworth/arsnouveau/common/entity/LightningEntity.java
@@ -41,7 +41,7 @@ public class LightningEntity extends LightningBoltEntity {
     @Nullable
     private ServerPlayerEntity caster;
 
-    public int amps;
+    public float amps;
     public int extendTimes;
 
     public float ampScalar;
